+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.80.0] - 2021-07-05
+
+### Added
+
+- RichErrorSummary to hydrate an ErrorSummary with the errors in a Play form.
+
+### Compatible with
+
+- [hmrc/hmrc-frontend v1.37.1](https://github.com/hmrc/hmrc-frontend/releases/tag/v1.37.1)
+- [hmrc/play-frontend-govuk v0.79.0](https://github.com/hmrc/play-frontend-govuk/releases/tag/v0.79.0)
+- [alphagov/govuk-frontend v3.13.0](https://github.com/alphagov/govuk-frontend/releases/tag/v3.13.0)
+
 ## [0.79.0] - 2021-06-29
 
 ### Fixed